+++ packages/seedless-onboarding-controller/src/types.ts
@@ -334,8 +334,9 @@ export type VaultData = {
   toprfAuthKeyPair: string;
   /**
    * The revoke token to revoke refresh token and get new refresh token and new revoke token.
+   * The revoke token may no longer be available after a large number of password changes. In this case, re-authentication is advised.
    */
-  revokeToken: string;
+  revokeToken?: string;
   /**
    * The access token used for pairing with profile sync auth service and to access other services.
    */